piper/pyproject.toml
[project]
name = "piper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blinker==1.9.0",
    "click==8.2.1",
    "coloredlogs==15.0.1",
    "flask==3.0.2",
    "flatbuffers==25.2.10",
    "humanfriendly==10.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "numpy>=1.20.0",
    "onnxruntime>=1.14.0",
    "packaging==25.0",
    "piper-phonemize==1.1.0",
    "piper-tts==1.2.0",
    "protobuf==6.31.1",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "striprtf==0.0.26",
    "sympy==1.14.0",
    "werkzeug==3.1.3",
]
